fix(cron): run LoTW sync in-process instead of self-fetching localhost
The unified sync cron invoked /api/lotw/upload and /api/lotw/download by
fetching NEXTAUTH_URL || http://localhost:3000 — in serverless deployments
nothing listens on localhost, so every LoTW leg failed with
ECONNREFUSED 127.0.0.1:3000.

Extract the upload/download logic into src/lib/lotw-sync.ts
(performLotwUpload / performLotwDownload) and call it directly from both
the cron and the API routes. The routes keep byte-identical response
shapes and status codes; the cron no longer depends on NEXTAUTH_URL.

With the self-fetch gone, the X-Cron-Job request mode on the LoTW routes
has no remaining callers, so it is removed — they now always require a
user token, and cron work is scoped by the station's own user_id inside
the lib.
