+++ src/app/api/lotw/download/route.ts
@@ -2,31 +2,15 @@
 
 import { NextRequest, NextResponse } from 'next/server';
 import { verifyToken } from '@/lib/auth';
-import { hasValidCronSecret } from '@/lib/cron-auth';
 import { query } from '@/lib/db';
-import { parseLoTWAdif, matchLoTWConfirmations, buildLoTWDownloadUrl, decryptString } from '@/lib/lotw';
-import { LotwDownloadRequest, LotwDownloadResponse, ContactWithLoTW } from '@/types/lotw';
+import { performLotwDownload } from '@/lib/lotw-sync';
+import { LotwDownloadRequest } from '@/types/lotw';
 
 export async function POST(request: NextRequest) {
   try {
-    // Cron mode requires the valid CRON_SECRET Bearer token — the X-Cron-Job
-    // header is only a mode discriminator and grants nothing by itself
-    // (it is spoofable by any caller).
-    const cronHeaderPresent = request.headers.get('X-Cron-Job') === 'true';
-    const isCronJob = cronHeaderPresent && hasValidCronSecret(request);
-    if (cronHeaderPresent && !isCronJob) {
+    const user = await verifyToken(request);
+    if (!user) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-    }
-    let user = null;
-
-    if (isCronJob) {
-      // For cron jobs, we'll get the user from the station_id
-      user = null; // Will be set later
-    } else {
-      user = await verifyToken(request);
-      if (!user) {
-        return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-      }
     }
 
     const body: LotwDownloadRequest = await request.json();
@@ -38,355 +22,15 @@ export async function POST(request: NextRequest) {
       }, { status: 400 });
     }
 
-    // Verify station exists and get LoTW credentials
-    let stationResult;
-    if (isCronJob) {
-      // For cron jobs, just verify station exists
-      stationResult = await query(
-        `SELECT s.id, s.callsign, s.lotw_username, s.lotw_password, s.user_id,
-                s.lotw_last_qsl_rcvd_date, u.third_party_services
-         FROM stations s
-         JOIN users u ON s.user_id = u.id
-         WHERE s.id = $1`,
-        [station_id]
-      );
-    } else {
-      // For regular requests, verify station belongs to user
-      if (!user) {
-        return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-      }
-      stationResult = await query(
-        `SELECT s.id, s.callsign, s.lotw_username, s.lotw_password, s.user_id,
-                s.lotw_last_qsl_rcvd_date, u.third_party_services
-         FROM stations s
-         JOIN users u ON s.user_id = u.id
-         WHERE s.id = $1 AND s.user_id = $2`,
-        [station_id, parseInt(user.userId)]
-      );
-    }
+    const result = await performLotwDownload({
+      stationId: station_id,
+      requesterUserId: parseInt(user.userId),
+      dateFrom: date_from,
+      dateTo: date_to,
+      downloadMethod: download_method,
+    });
 
-    if (stationResult.rows.length === 0) {
-      return NextResponse.json({
-        error: 'Station not found or access denied'
-      }, { status: 404 });
-    }
-
-    const station = stationResult.rows[0];
-    const userId = isCronJob ? station.user_id : parseInt(user!.userId);
-
-    // Get LoTW credentials from station or user third_party_services
-    let lotwUsername = station.lotw_username;
-    let lotwPassword = station.lotw_password;
-    let credentialSource = 'none';
-
-    // Try user's third_party_services if station doesn't have credentials
-    if (!lotwUsername || !lotwPassword) {
-      const thirdPartyServices = station.third_party_services;
-
-      if (thirdPartyServices?.lotw) {
-        lotwUsername = thirdPartyServices.lotw.username;
-        try {
-          lotwPassword = decryptString(thirdPartyServices.lotw.password);
-          credentialSource = 'user';
-        } catch (error) {
-          console.error('[LoTW Download] Failed to decrypt user password:', error);
-          lotwPassword = undefined;
-        }
-      }
-    } else if (lotwPassword) {
-      // Decrypt station password if it exists
-      try {
-        lotwPassword = decryptString(lotwPassword);
-        credentialSource = 'station';
-      } catch (error) {
-        console.error('[LoTW Download] Failed to decrypt station password:', error);
-        lotwPassword = undefined;
-      }
-    }
-
-    if (!lotwUsername || !lotwPassword) {
-      console.error(`[LoTW Download] No valid credentials found. Username: ${!!lotwUsername}, Password: ${!!lotwPassword}, Source attempted: ${credentialSource}`);
-
-      const errorMessage = credentialSource === 'none'
-        ? 'LoTW credentials not configured. Please configure LoTW credentials in Settings > LoTW Integration (either at User level for all stations, or Station level for this specific station).'
-        : 'LoTW credentials are incomplete or invalid. Please reconfigure them in Settings > LoTW Integration.';
-
-      return NextResponse.json({ error: errorMessage }, { status: 400 });
-    }
-
-    // Create download log entry
-    const downloadLogResult = await query(
-      `INSERT INTO lotw_download_logs 
-       (station_id, user_id, date_from, date_to, status, download_method) 
-       VALUES ($1, $2, $3, $4, 'pending', $5) 
-       RETURNING id`,
-      [station_id, userId, date_from || null, date_to || null, download_method]
-    );
-
-    const downloadLogId = downloadLogResult.rows[0].id;
-
-    try {
-      // Update status to processing
-      await query(
-        'UPDATE lotw_download_logs SET status = $1, started_at = NOW() WHERE id = $2',
-        ['processing', downloadLogId]
-      );
-
-      // Incremental fetch for scheduled runs: resume from the newest QSL date
-      // already applied (qso_qslsince filters by QSL date, not QSO date, so
-      // the local contacts query below must stay unfiltered by this date —
-      // confirmations can arrive for arbitrarily old QSOs).
-      let urlDateFrom = date_from;
-      if (!urlDateFrom && download_method !== 'manual' && station.lotw_last_qsl_rcvd_date) {
-        const last = station.lotw_last_qsl_rcvd_date;
-        urlDateFrom = last instanceof Date ? last.toISOString().split('T')[0] : String(last);
-      }
-
-      // Build LoTW download URL
-      const downloadUrl = buildLoTWDownloadUrl(lotwUsername, lotwPassword, {
-        dateFrom: urlDateFrom,
-        dateTo: date_to,
-        ownCallsign: station.callsign,
-      });
-
-      // Download confirmations from LoTW
-      let adifContent: string;
-      try {
-        const downloadResponse = await fetch(downloadUrl, {
-          method: 'GET',
-          headers: {
-            'User-Agent': 'Nextlog/1.0.0',
-          },
-        });
-
-        if (!downloadResponse.ok) {
-          throw new Error(`LoTW download failed: ${downloadResponse.status} ${downloadResponse.statusText}`);
-        }
-
-        adifContent = await downloadResponse.text();
-
-        // Check if the response indicates an authentication error
-        if (adifContent.includes('Invalid login') || adifContent.includes('Login failed')) {
-          throw new Error('Invalid LoTW credentials');
-        }
-
-      } catch (downloadError) {
-        console.error('LoTW download error:', downloadError);
-
-        await query(
-          `UPDATE lotw_download_logs 
-           SET status = 'failed', completed_at = NOW(), 
-               error_message = $1 
-           WHERE id = $2`,
-          [`Download from LoTW failed: ${downloadError instanceof Error ? downloadError.message : 'Unknown error'}`, downloadLogId]
-        );
-
-        return NextResponse.json({
-          success: false,
-          download_log_id: downloadLogId,
-          error_message: `Download from LoTW failed: ${downloadError instanceof Error ? downloadError.message : 'Unknown error'}`
-        }, { status: 500 });
-      }
-
-      // Parse ADIF confirmations
-      const confirmations = parseLoTWAdif(adifContent);
-
-      if (confirmations.length === 0) {
-        await query(
-          `UPDATE lotw_download_logs 
-           SET status = 'completed', completed_at = NOW(), 
-               qso_count = 0, confirmations_found = 0, confirmations_matched = 0,
-               error_message = 'No confirmations found in LoTW response' 
-           WHERE id = $1`,
-          [downloadLogId]
-        );
-
-        const response: LotwDownloadResponse = {
-          success: true,
-          download_log_id: downloadLogId,
-          confirmations_found: 0,
-          confirmations_matched: 0,
-          confirmations_unmatched: 0,
-          error_message: 'No confirmations found in LoTW response'
-        };
-
-        return NextResponse.json(response);
-      }
-
-      // Get contacts from this station to match against. Join the station
-      // callsign so the matcher can disambiguate multi-station accounts
-      // against LoTW's app_lotw_owncall.
-      let contactQuery = `
-        SELECT c.*, s.callsign as station_callsign
-        FROM contacts c
-        JOIN stations s ON c.station_id = s.id
-        WHERE c.user_id = $1 AND c.station_id = $2
-      `;
-      const queryParams: (string | number)[] = [userId, station_id];
-
-      if (date_from || date_to) {
-        if (date_from) {
-          contactQuery += ` AND c.datetime >= $3`;
-          queryParams.push(date_from);
-        }
-        if (date_to) {
-          const paramIndex = queryParams.length + 1;
-          contactQuery += ` AND c.datetime <= $${paramIndex}`;
-          queryParams.push(date_to);
-        }
-      }
-
-      contactQuery += ` ORDER BY c.datetime ASC`;
-
-      const contactsResult = await query(contactQuery, queryParams);
-      const contacts: ContactWithLoTW[] = contactsResult.rows;
-
-      // Match confirmations with local contacts
-      const matches = matchLoTWConfirmations(confirmations, contacts);
-
-      let matchedCount = 0;
-      let unmatchedCount = confirmations.length;
-      let failedCount = 0;
-      let firstApplyError: string | null = null;
-      const appliedQslDates: string[] = [];
-
-      // Apply each confirmation: set LoTW QSL flags, enrich location fields
-      // (state/county/CQZ/ITUZ/DXCC/country/grid/iota), and cross-flag QRZ
-      // for re-upload when the QSO was already in QRZ ('Y' → 'M').
-      for (const match of matches) {
-        const conf = match.confirmation;
-
-        // Build a dynamic UPDATE — only set fields LoTW returned non-empty.
-        const sets: string[] = [
-          'qsl_lotw = true',
-          `lotw_qsl_rcvd = 'Y'`,
-          'lotw_match_status = $1',
-          'updated_at = NOW()',
-        ];
-        const params: (string | number | Date)[] = [match.matchStatus];
-
-        const addEnrich = (col: string, value: string | undefined) => {
-          if (!value) return;
-          params.push(value);
-          sets.push(`${col} = $${params.length}`);
-        };
-        addEnrich('state', conf.state);
-        addEnrich('cnty', conf.cnty);
-        if (conf.cqz) {
-          const n = parseInt(conf.cqz, 10);
-          if (!Number.isNaN(n)) { params.push(n); sets.push(`cqz = $${params.length}`); }
-        }
-        if (conf.ituz) {
-          const n = parseInt(conf.ituz, 10);
-          if (!Number.isNaN(n)) { params.push(n); sets.push(`ituz = $${params.length}`); }
-        }
-        if (conf.dxcc) {
-          const n = parseInt(conf.dxcc, 10);
-          if (!Number.isNaN(n)) { params.push(n); sets.push(`dxcc = $${params.length}`); }
-        }
-        addEnrich('country', conf.country);
-        // Only update gridsquare if LoTW reported one and ours was missing/different.
-        if (conf.gridsquare && conf.gridsquare !== match.contact.grid_locator) {
-          params.push(conf.gridsquare);
-          sets.push(`grid_locator = $${params.length}`);
-        }
-        // Cross-sync: if QRZ already shipped this QSO, mark for re-upload so the
-        // updated lotw_qsl_rcvd / location fields propagate. Wavelog's 'M' flag.
-        if (match.contact.qrz_qsl_sent === 'Y') {
-          sets.push(`qrz_qsl_sent = 'M'`);
-        }
-        // qsl_lotw_date is assigned exactly once — Postgres rejects an UPDATE
-        // that sets the same column twice. Prefer LoTW's own QSL date
-        // (YYYY-MM-DD already); fall back to today (UTC) so the applied date
-        // still advances the incremental-download watermark below.
-        const qslDate = conf.qsl_rcvd_date || new Date().toISOString().slice(0, 10);
-        params.push(qslDate);
-        sets.push(`qsl_lotw_date = $${params.length}::date`);
-
-        params.push(match.contact.id);
-        try {
-          await query(
-            `UPDATE contacts SET ${sets.join(', ')} WHERE id = $${params.length}`,
-            params
-          );
-          matchedCount++;
-          unmatchedCount--;
-          appliedQslDates.push(qslDate);
-        } catch (applyError) {
-          // Isolate per-row failures so one bad row can't discard the batch.
-          failedCount++;
-          const msg = applyError instanceof Error ? applyError.message : 'Unknown error';
-          if (!firstApplyError) firstApplyError = msg;
-          console.error(
-            `[LoTW Download] Failed to apply confirmation to contact ${match.contact.id}:`,
-            applyError
-          );
-        }
-      }
-
-      // Persist the most recent qsl_rcvd_date so the next download can use it
-      // as qso_qslsince for an incremental fetch. Only consider confirmations
-      // that actually applied — advancing past a failed row would skip it on
-      // the next incremental run.
-      if (appliedQslDates.length > 0) {
-        const maxDate = appliedQslDates
-          .sort()
-          .pop();
-        if (maxDate) {
-          await query(
-            `UPDATE stations SET lotw_last_qsl_rcvd_date = GREATEST(
-                COALESCE(lotw_last_qsl_rcvd_date, '1970-01-01'::date), $1::date
-             ) WHERE id = $2`,
-            [maxDate, station_id]
-          );
-        }
-      }
-
-      // Update download log as completed; surface partial apply failures so
-      // they are visible in the sync history instead of silently dropped.
-      const applyErrorMessage = failedCount > 0
-        ? `${failedCount} matched confirmation(s) failed to apply: ${firstApplyError}`
-        : null;
-
-      await query(
-        `UPDATE lotw_download_logs
-         SET status = 'completed', completed_at = NOW(),
-             qso_count = $1, confirmations_found = $2,
-             confirmations_matched = $3, confirmations_unmatched = $4,
-             error_message = $5
-         WHERE id = $6`,
-        [contacts.length, confirmations.length, matchedCount, unmatchedCount, applyErrorMessage, downloadLogId]
-      );
-
-      const response: LotwDownloadResponse = {
-        success: true,
-        download_log_id: downloadLogId,
-        confirmations_found: confirmations.length,
-        confirmations_matched: matchedCount,
-        confirmations_unmatched: unmatchedCount,
-        ...(applyErrorMessage ? { error_message: applyErrorMessage } : {})
-      };
-
-      return NextResponse.json(response);
-
-    } catch (error) {
-      console.error('Download processing error:', error);
-
-      // Update log with error
-      await query(
-        `UPDATE lotw_download_logs 
-         SET status = 'failed', completed_at = NOW(), 
-             error_message = $1 
-         WHERE id = $2`,
-        [error instanceof Error ? error.message : 'Unknown error', downloadLogId]
-      );
-
-      return NextResponse.json({
-        success: false,
-        download_log_id: downloadLogId,
-        error_message: error instanceof Error ? error.message : 'Unknown error'
-      }, { status: 500 });
-    }
+    return NextResponse.json(result.body, { status: result.status });
 
   } catch (error) {
     console.error('LoTW download error:', error);
